+++ GISChat/Properties/AssemblyInfo.cs
@@ -6,6 +6,6 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyCompany("Zsolt Zoltán Fehér - GIS Expert")]
 [assembly: AssemblyProduct("GIS Chat")]
 [assembly: AssemblyCopyright("Copyright © 2026 Zsolt Zoltán Fehér. MIT License.")]
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
+[assembly: AssemblyVersion("1.1.0.0")]
+[assembly: AssemblyFileVersion("1.1.0.0")]
 [assembly: ComVisible(false)]